refactor: move PendingData to gateway types
