+++ crates/gateway-types/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod class_hash;
 pub mod error;
+pub mod pending;
 pub mod reply;
 pub mod request;
