+++ crates/rpc/src/v02.rs
@@ -1,10 +1,10 @@
 use super::error::RpcError;
 use crate::cairo::ext_py;
 use crate::gas_price;
-use crate::state::PendingData;
 use crate::state::SyncState;
 use pathfinder_common::ChainId;
 use pathfinder_storage::Storage;
+use starknet_gateway_types::pending::PendingData;
 use std::sync::Arc;
 
 mod common;
